Se generan/eliminan ordenes con verificacion

+++ Proyecto_Final/EntregaFinal/02_controllers/orders.controllers.js
@@ -26,9 +26,16 @@ const postOrder = async(req, res) => {  // Devuelve el ID de la orden agregada
     let getCart = await carritos.getById(parseInt(req.params.idCart))
     let data = []
     let count = 0
+    let emailToken = req.email
 
-    if (getCart.length === 0) {
+    if (getCart.length === 0) {//Chequeo si existe el carrito
         return res.send({ error : `No existe carrito con ID=${parseInt(req.params.idCart)} `})
+    }
+    if (getCart[0].productos.length === 0){////Chequeo que el carrito contenga productos
+        return res.send({ error : `No se puede generar una Orden con el carrito vacio`})
+    }
+    if(getCart[0].email !== emailToken){//Chequeo si el usuario tiene acceso a ese carrito
+        return res.send({ error : `El usuario ${emailToken} no tiene un carrito con ID=${parseInt(req.params.idCart)}`})
     }else{
         for(let i = 0; i < getCart[0].productos.length; i++){
             let auxObj = {
@@ -54,11 +61,24 @@ const postOrder = async(req, res) => {  // Devuelve el ID de la orden agregada
         })
         return res.send(idOrder)               
     }
+    
+    
 }
 
 const deleteOrderById = async(req, res) => {
-    let deleteiDOrder = await ordenes.deleteById(parseInt(req.params.id))   
-    return res.send(deleteiDOrder)     
+    let getOrder = await ordenes.getById(parseInt(req.params.id)) 
+    let emailToken = req.email 
+
+    if (getOrder.length === 0) {//Chequeo si existe el carrito
+        return res.send({ error : `No existe orden con ID=${parseInt(req.params.id)} `})        
+    }
+    
+    if(getOrder[0].email !== emailToken){//Chequeo si el usuario tiene acceso a esa orden
+        return res.send({ error : `El usuario ${emailToken} no tiene una orden con ID=${parseInt(req.params.id)} `})
+    }else{
+        let deleteiDOrder = await ordenes.deleteById(parseInt(req.params.id))   
+        return res.send(deleteiDOrder)        
+    }    
 }
 
 export { getOrders, getOrderById, postOrder, deleteOrderById }
